Dhruva-Platform-1/server/requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==1.0.0
pymongo==3.12.0
motor==2.5.1
pydantic<2.0.0,>=1.10.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==3.2.0
requests==2.31.0
aiohttp==3.8.1
websockets==10.0
redis==4.1.0
redis-om==0.1.2
celery==5.2.3
flower==1.0.0
prometheus-client==0.11.0
jsonpickle==3.0.2
psycopg2-binary==2.9.9
SQLAlchemy==1.4.41
fastapi-sqlalchemy==0.2.1
soundfile==0.12.1
email-validator==2.0.0
PyJWT==2.8.0
pydub==0.25.1
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
librosa==0.10.1
torch==2.0.1
torchaudio==2.0.2
transformers==4.30.2
argon2-cffi==21.3.0
gevent==23.9.1
httpx==0.28.1
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1
black==21.9b0
isort==5.9.3
flake8==3.9.2
mypy==0.910
types-requests==2.26.0
types-python-jose==3.3.0
types-redis==4.1.0
types-urllib3==1.26.0
sqlalchemy-timescaledb==0.4.1